change column of numeroHora to double with two decimals

+++ database/migrations/2023_12_12_220550_create_jornadas_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->text('descripcion')->nullable();
             $table->time('horaInicial');
             $table->time('horaFinal');
-            $table->integer('numeroHoras');
+            $table->double('numeroHoras', 8, 2);
             $table->unsignedInteger('idCompany');
             $table->foreign('idCompany')->references('id')->on('company');
             $table->timestamps();
